[#75] Refactor box test reports structure
This is preparation before introducing reports for MicroPython box
tests. A more formal structure is introduced, so that current reports
sit in `python/box.tests/reports/pythonTest`, and the last part of this
path is actually the name of the Gradle task, which in case of
MicroPython will be `microPythonTest`.