[#94] Remove unreferenced code
Anything that was marked as unused in IntelliJ was removed. The
analysis was very shallow, and didn't check e.g. if actually a lowering
is used. It's in scope of the next changes.

The only exceptions are some entities which symbols are not used in
compile time, but they are somehow used in runtime. Each such
occurrence was marked with a comment.